+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Uptime Robot plugin for Craft CMS 3.x Changelog
 
+## 1.1.0 - 2018-12-14
+### Fixed
+- Better exception handling (Uptime Robot maintenance downtime for instance)
+
+### Added
+- Average response time info and response times graph
+
+### Changed
+- Redesigned monitor view
+
+
 ## 1.0.0 - 2018-10-24
 ### Added
 - Initial release